requirements.txt
flask
pyserial
flask-login
gunicorn
